sockets lib renamed
