+++ src/main/js/app/components/dashboard/Dashboard.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from 'styled-components';
 import { List, Label } from 'semantic-ui-react';
 import { history } from 'app/app';
-import * as sockets from 'socketsLib';
+import * as sockets from 'sockets';
 import { routesList } from 'app/lib/selectOptions';
 
 const DashboardContainer = styled.div`
