+++ src/main/js/app/components/users/users-list/UsersList.jsx
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 import { Label, List } from 'semantic-ui-react';
 import { ItemContent } from '../../dashboard/Dashboard';
 import PaginatorList from 'components/common/paginator-list/PaginatorList';
-import * as sockets from 'socketsLib';
+import * as sockets from 'sockets';
 
 const ListItem = ({ item }) => (
     <Link to={`/users/${item.hid}`} replace>
