+++ src/main/js/app/lib/sockets/dashboard.js
